+++ .github/workflows/ci_release.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - main
+      - ci/ramin/integration-tests
     # Trigger on version tags
     tags:
       - "v*"
