+++ nodebuilder/tests/api_test.go
@@ -1,3 +1,5 @@
+//go:build api || integration
+
 package tests
 
 import (
@@ -13,26 +15,10 @@ import (
 	"github.com/celestiaorg/celestia-node/api/rpc/client"
 	"github.com/celestiaorg/celestia-node/blob"
 	"github.com/celestiaorg/celestia-node/blob/blobtest"
-	"github.com/celestiaorg/celestia-node/libs/authtoken"
 	"github.com/celestiaorg/celestia-node/nodebuilder"
 	"github.com/celestiaorg/celestia-node/nodebuilder/node"
 	"github.com/celestiaorg/celestia-node/nodebuilder/tests/swamp"
 )
-
-func getAdminClient(ctx context.Context, nd *nodebuilder.Node, t *testing.T) *client.Client {
-	t.Helper()
-
-	signer := nd.AdminSigner
-	listenAddr := "ws://" + nd.RPCServer.ListenAddr()
-
-	jwt, err := authtoken.NewSignedJWT(signer, []auth.Permission{"public", "read", "write", "admin"})
-	require.NoError(t, err)
-
-	client, err := client.NewClient(ctx, listenAddr, jwt)
-	require.NoError(t, err)
-
-	return client
-}
 
 func TestNodeModule(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), swamp.DefaultTestTimeout)
