+++ nodebuilder/tests/p2p_test.go
@@ -1,3 +1,5 @@
+//go:build p2p || integration
+
 package tests
 
 import (
